style: cleanup of variable names
